Fixed bug
Fix for pre #1948

Fix bytes order for SHA-1 output

Changed logic guessed_protocol.
We use guessed_protocol_id as app_protocol if it is not excluded.
If guessed_protocol_id_by_ip is found, it will be used as
app_protocol if not already defined, otherwise it can be used as
master_protocol if not already defined.